Add compress method for Genie
